Layout with some styling and some more SocketIO
I have made the index to be like a start page where all the users online will be listed. You will be listed over there for everyone else, as soon as you enter your name/alias whatever in the input field to the right.

In the making are a field where you can see the username of the user you have marked, and then there will be probably a couple more buttons, one to send match invite, another to accept invites others have send to you.