Implemented Challenging system.
It is now possible to mark the player you want to challenge, and click challenge. It will then pop up in the Table below the Challenge button.
A timer will start and count down 60 seconds, that's how long you have to accept or deny a challenge before it dissaperes again.