Implemented click function for the game tiles.
It now takes the data-col and data-row and your opponents id and sends to the server.
Needed implementing is that the server chooses which player that is red and which is black. After that I need a boolean on the 'TicTacToe' class, to check whether it is your turn or not.