Most game logic done.
Need to make the last diagonal check from bottom left to top right.

After that is done I will start making it a bit prettier, and then I will make it so Your oppenent is listed on the side, And I will make the win/loss screen with a button that directs one back to the start page.