Visual update.
It now contains the pics for X and O and uses those instead of color.

I have made a game infoo field in the left side underneath user input.
This shall contain opponents name, and then who's turn it is. It shall also contain the name of the winner in big green text, and then display a exit match button in the buttom of the left side panel.