Changing the css styling once again, because it did not help on phones.
I am making a button in the left bottom corner, which is going to be the change to mobiel version button.

So far it makes the title text larger, and then it changes the userlist class, so when a div inside the userlist_container is clicked it changes the classes again to show the info.

Not the best and most functioning way to make it work better on phones. But I have to start learning it somehow and somewhere.