Undid most of the css styling changes.
I did this because it messed up the game, and I can't really keep track of all the components anymore.

I still have the css styling class for mobiles, and I still have the javascript file that handles the changing of the classes.

But for some reason it stopped working with the server, And I forgot to keep track of when it stopped working. So I undid most, and I will maybe implement a button that changes the layout so it is more mobile friendly. But for now, it is a too big of a task, in the middle of exams, to be start learning this. Plus I still think this is the wrong approach to this.